fix:正火去氢记录相关操作增加权限限制

+++ mes-modules/produce/src/main/java/com/richfit/mes/produce/controller/NormalizeDehydroRecordController.java
@@ -9,11 +9,14 @@ import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.mysql.cj.util.StringUtils;
 import com.richfit.mes.common.core.api.CommonResult;
+import com.richfit.mes.common.core.api.ResultCode;
 import com.richfit.mes.common.core.base.BaseController;
+import com.richfit.mes.common.core.exception.GlobalException;
 import com.richfit.mes.common.model.base.Device;
 import com.richfit.mes.common.model.base.DevicePerson;
 import com.richfit.mes.common.model.base.SequenceSite;
 import com.richfit.mes.common.model.produce.*;
+import com.richfit.mes.common.model.sys.TenantUser;
 import com.richfit.mes.common.model.sys.vo.TenantUserVo;
 import com.richfit.mes.common.model.util.DrawingNoUtil;
 import com.richfit.mes.common.model.util.OrderUtil;
@@ -39,6 +42,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
  * @author 胡甲
@@ -54,11 +58,13 @@ public class NormalizeDehydroRecordController extends BaseController {
     private NormalizeDehydroRecordService normalizeDehydroRecordService;
     @Autowired
     private PrechargeFurnaceService prechargeFurnaceService;
-
+    @Autowired
+    private SystemServiceClient systemServiceClient;
 
     @ApiOperation(value = "正火去氢工序控制查询", notes = "正火去氢工序控制查询")
     @PostMapping("/pageNormalizeDehydroRecord")
     public CommonResult<IPage<NormalizeDehydroRecord>> pageNormalizeDehydroRecord(@RequestBody NormalizeDehydroRecord normalizeDehydroRecord) {
+        TenantUserDetails currentUser = SecurityUtils.getCurrentUser();
         QueryWrapper<NormalizeDehydroRecord> queryWrapper = new QueryWrapper<>();
 
         if(!StringUtils.isNullOrEmpty(normalizeDehydroRecord.getSerialNo())){
@@ -72,6 +78,21 @@ public class NormalizeDehydroRecordController extends BaseController {
         }
         if (!StringUtils.isNullOrEmpty(normalizeDehydroRecord.getEndTime())) {
             queryWrapper.le("create_time", normalizeDehydroRecord.getEndTime() + " 23:59:59");
+        }
+        boolean isBzz = normalizeDehydroRecordService.isBzz(currentUser);
+        //班组长查询
+        if (isBzz) {
+            //班组长查询同班员工号
+            List<TenantUser> tenantUserList = systemServiceClient.queryClass(currentUser.getUsername());
+            if(CollectionUtils.isNotEmpty(tenantUserList)){
+                List<String> userIdList = tenantUserList.stream().map(TenantUser::getUserAccount).collect(Collectors.toList());
+                userIdList.add(currentUser.getUsername());
+                queryWrapper.in("create_by",userIdList);
+            }
+        }
+        //普通操作工查询
+        else {
+            queryWrapper.eq("create_by",currentUser.getUsername());
         }
         return CommonResult.success(normalizeDehydroRecordService.page(new Page<>(normalizeDehydroRecord.getPage(),normalizeDehydroRecord.getLimit()), queryWrapper));
     }
@@ -110,7 +131,14 @@ public class NormalizeDehydroRecordController extends BaseController {
     @ApiOperation(value = "正火去氢工序控制删除", notes = "正火去氢工序控制删除")
     @PostMapping("/deleteNormalizeDehydroRecord")
     public CommonResult<Boolean> deleteNormalizeDehydroRecord(@RequestBody List<String> idList) {
-        return CommonResult.success(normalizeDehydroRecordService.removeByIds(idList));
+        TenantUserDetails currentUser = SecurityUtils.getCurrentUser();
+        //只有班组长有删除权限
+        if(normalizeDehydroRecordService.isBzz(currentUser)){
+            return CommonResult.success(normalizeDehydroRecordService.removeByIds(idList));
+        }else {
+            throw new GlobalException("删除失败!无班组长权限", ResultCode.FAILED);
+        }
+
     }
 
     @ApiOperation(value = "正火去氢工序控审核", notes = "正火去氢工序控审核")
@@ -118,6 +146,10 @@ public class NormalizeDehydroRecordController extends BaseController {
     public CommonResult<Boolean> auditNormalizeDehydroRecord(@ApiParam(value = "idList") @RequestBody List<String> idList,
                                                              @ApiParam(value = "审核状态  0 未审核  1 通过,2 未通过",required = true)@RequestParam Integer status) {
         TenantUserDetails currentUser = SecurityUtils.getCurrentUser();
+        //只有班组长有审核权限
+        if(!normalizeDehydroRecordService.isBzz(currentUser)){
+            throw new GlobalException("审核失败!无班组长权限",ResultCode.FAILED);
+        }
         UpdateWrapper<NormalizeDehydroRecord> updateWrapper=new UpdateWrapper<>();
         updateWrapper.in("id",idList);
         updateWrapper.set("audit_status",status);
